refactor(6/introduce): SecondInteraction 리팩토링
